Revert "Converted experience to long, updated the migrator"
This reverts commit 74585f634d96fad6577e1c21a88573cbf4cf53ee.


Former-commit-id: fd025a8dc83dd9a6c05eb8e7a25ff8a3dcefc332

+++ Intersect Editor/Forms/Editors/frmClass.cs
@@ -877,7 +877,7 @@ namespace Intersect.Editor.Forms
 
         private void nudExpIncrease_ValueChanged(object sender, EventArgs e)
         {
-            _editorItem.ExpIncrease = (long) nudExpIncrease.Value;
+            _editorItem.ExpIncrease = (int) nudExpIncrease.Value;
         }
 
         private void nudHpIncrease_ValueChanged(object sender, EventArgs e)
@@ -933,7 +933,7 @@ namespace Intersect.Editor.Forms
 
         private void nudBaseExp_ValueChanged(object sender, EventArgs e)
         {
-            _editorItem.BaseExp = (long) nudBaseExp.Value;
+            _editorItem.BaseExp = (int) nudBaseExp.Value;
         }
 
         private void nudBaseHP_ValueChanged(object sender, EventArgs e)
